Use device time for authentication (#9)
Many cameras have wrong time and/or timezone and often end up rejecting
client authentication due to their time skew. With this change, we first
ask the device for its own time (an endpoint that is itself not
authenticated) and then use the computed time skew in any further
request authentication.

Fixes: https://eyetowers.atlassian.net/browse/SW-97.